MDEV-21652 drop SYS_FOREIGN[_COLS] starter
- fk_create_legacy_storage debug flag to create SYS_FOREIGN[_COLS] on
  innodb_eval_sql.

- fk_release_locks() iterates through all indexes and pages and releases
locks on SUPREMUM records placed there previously by SELECT FOR
UPDATE. Also it releases all table locks.

FIXME:
- row_purge_parse_undo_rec() acquires SYS_FOREIGN table reference but
  does not frees it
- don't create SYS_FOREGIN[_COLS] on startup
- foreign_key_legacy test
- FIXMEs in code

+++ storage/innobase/row/row0mysql.cc
@@ -3294,6 +3294,7 @@ dberr_t
 row_drop_table_check_legacy_fk(trx_t* trx, const char* table_name,
 			       row_drop_table_check_legacy_data& d)
 {
+	ut_ad(DB_SUCCESS == fk_legacy_storage_exists(false));
 	static const char sql_check[]
 		= "PROCEDURE FK_PROC () IS\n"
 		  "DECLARE FUNCTION row_drop_table_check_legacy_step;\n"
@@ -3323,6 +3324,10 @@ row_drop_table_check_legacy_fk(trx_t* trx, const char* table_name,
 
 	return DB_SUCCESS;
 }
+
+/** Drop SYS_FOREIGN[_COLS] tables if they are empty */
+dberr_t
+fk_cleanup_legacy_storage(bool lock_dict_mutex, trx_t* trx);
 #endif /* WITH_INNODB_LEGACY_FOREIGN_STORAGE */
 
 /** Drop a table for MySQL.
@@ -3519,30 +3524,39 @@ row_drop_table_for_mysql(
 #ifdef WITH_INNODB_LEGACY_FOREIGN_STORAGE
 		row_drop_table_check_legacy_data data;
 
-		err = row_drop_table_check_legacy_fk(trx, table->name.m_name,
-						     data);
-		if (err != DB_SUCCESS) {
+		err = fk_legacy_storage_exists(false);
+		if (err == DB_CORRUPTION) {
 			goto funct_exit;
 		}
-		if (data.found) {
-			FILE* ef = dict_foreign_err_file;
+		if (err == DB_TABLE_NOT_FOUND) {
+			err = DB_SUCCESS;
+		} else {
+			ut_ad(err == DB_SUCCESS);
+			err = row_drop_table_check_legacy_fk(
+				trx, table->name.m_name, data);
+			if (err != DB_SUCCESS) {
+				goto funct_exit;
+			}
+			if (data.found) {
+				FILE* ef = dict_foreign_err_file;
 
-			err = DB_CANNOT_DROP_CONSTRAINT;
+				err = DB_CANNOT_DROP_CONSTRAINT;
 
-			mutex_enter(&dict_foreign_err_mutex);
-			rewind(ef);
-			ut_print_timestamp(ef);
+				mutex_enter(&dict_foreign_err_mutex);
+				rewind(ef);
+				ut_print_timestamp(ef);
 
-			fputs("  Cannot drop table ", ef);
-			ut_print_name(ef, trx, name);
-			fputs("\n"
-			      "because it is referenced by ",
-			      ef);
-			ut_print_name(ef, trx, data.foreign_name);
-			putc('\n', ef);
-			mutex_exit(&dict_foreign_err_mutex);
+				fputs("  Cannot drop table ", ef);
+				ut_print_name(ef, trx, name);
+				fputs("\n"
+				"because it is referenced by ",
+				ef);
+				ut_print_name(ef, trx, data.foreign_name);
+				putc('\n', ef);
+				mutex_exit(&dict_foreign_err_mutex);
 
-			goto funct_exit;
+				goto funct_exit;
+			}
 		}
 #endif	  /* WITH_INNODB_LEGACY_FOREIGN_STORAGE */
 	} // if (!srv_read_only_mode && trx->check_foreigns)
@@ -3598,8 +3612,7 @@ defer:
 	shouldn't have to. There should never be record locks on a table
 	that is going to be dropped. */
 
-	if (table->get_ref_count() > 0 || table->n_rec_locks > 0
-	    || lock_table_has_locks(table)) {
+	if (table->get_ref_count() > 0 || lock_table_has_locks(table)) {
 		goto defer;
 	}
 
@@ -3658,8 +3671,7 @@ defer:
 #ifdef WITH_INNODB_LEGACY_FOREIGN_STORAGE
 	if (sqlcom != SQLCOM_TRUNCATE
 	    && strchr(name, '/')
-	    && dict_table_get_low("SYS_FOREIGN")
-	    && dict_table_get_low("SYS_FOREIGN_COLS")) {
+	    && (DB_SUCCESS == fk_legacy_storage_exists(false))) {
 		err = que_eval_sql(
 			info,
 			"PROCEDURE DROP_FOREIGN_PROC () IS\n"
@@ -3681,6 +3693,10 @@ defer:
 			"END LOOP;\n"
 			"CLOSE fk;\n"
 			"END;\n", FALSE, trx);
+		if (err != DB_SUCCESS) {
+			goto error;
+		}
+		err = fk_cleanup_legacy_storage(false, trx);
 		if (err != DB_SUCCESS) {
 			goto error;
 		}
@@ -4069,6 +4085,7 @@ row_delete_constraint_low(
 	const char*	id,		/*!< in: constraint id */
 	trx_t*		trx)		/*!< in: transaction handle */
 {
+	ut_ad(DB_SUCCESS == fk_legacy_storage_exists(false));
 	pars_info_t*	info = pars_info_create();
 
 	pars_info_add_str_literal(info, "id", id);
@@ -4112,6 +4129,11 @@ row_delete_constraint(
 		err = row_delete_constraint_low(id, trx);
 	}
 
+	if (err != DB_SUCCESS) {
+		return err;
+	}
+
+	err = fk_cleanup_legacy_storage(false, trx);
 	return(err);
 }
 #endif /* WITH_INNODB_LEGACY_FOREIGN_STORAGE */
@@ -4336,6 +4358,15 @@ row_rename_table_for_mysql(
 	}
 
 #ifdef WITH_INNODB_LEGACY_FOREIGN_STORAGE
+	err = fk_legacy_storage_exists(false);
+	if (err == DB_CORRUPTION) {
+		goto end;
+	}
+	if (err == DB_TABLE_NOT_FOUND) {
+		err = DB_SUCCESS;
+		goto fk_skip_rename;
+	}
+
 	if (!new_is_tmp) {
 		/* Rename all constraints. */
 		char	new_table_name[MAX_TABLE_NAME_LEN + 1];
@@ -4470,6 +4501,7 @@ row_rename_table_for_mysql(
 	if (err != DB_SUCCESS) {
 		goto end;
 	}
+fk_skip_rename:
 #endif /* WITH_INNODB_LEGACY_FOREIGN_STORAGE */
 
 	if ((dict_table_has_fts_index(table)
